move level ui to seperate module
